chore(chargePool): split 892-line service into focused modules
- Extract types to chargePool/types.ts
- Extract shared helpers to chargePool/helpers.ts
- Split into chargePoolConsume, chargePoolRecover, chargePoolPreview, chargePoolSync
- Update all consumers to use #utils/chargePool/* path aliases
- Update Shared Code Inventory in STYLE_GUIDE.md
