+++ docs/STYLE_GUIDE.md
@@ -1397,7 +1397,7 @@ export const myStore = writable<MyStoreState>(initialState);
 | `combatManaRules` | `src/utils/combatManaRules.ts` | Combat mana calculation and rules |
 | `itemSourceRules` | `src/utils/itemSourceRules.ts` | Resolve compendium source IDs and rules for embedded items |
 | `ChargePoolRuleConfig` | `src/utils/chargePoolRuleConfig.ts` | Shared charge-system constants (scopes, triggers, flags) |
-| `ChargePoolService` | `src/utils/chargePoolService.ts` | Charge pool sync, consumption, and recovery operations |
+| `ChargePoolService` | `src/utils/chargePool/chargePoolConsume.ts`, `chargePoolRecover.ts`, `chargePoolPreview.ts`, `chargePoolSync.ts` | Charge pool sync, consumption, and recovery operations |
 | `manaRecovery` | `src/utils/manaRecovery.ts` | Mana recovery calculations |
 | `prelocalize()` | `src/utils/prelocalize.ts` | Pre-localize configuration objects |
 | `getChoicesFromCompendium()` | `src/utils/getChoicesFromCompendium.ts` | Fetch selectable options from compendiums |
