+++ src/view/sheets/pages/PlayerCharacterInventoryTab.svelte
@@ -3,7 +3,7 @@
 	import type PlayerCharacterSheet from '#documents/sheets/PlayerCharacterSheet.svelte.js';
 	import { RulesManager } from '#managers/RulesManager.js';
 	import localize from '#utils/localize.js';
-	import { ChargePoolService } from '#utils/chargePoolService.js';
+	import { getPools, getPoolsForItem } from '#utils/chargePool/chargePoolSync.js';
 	import shouldFlashDroppedItem from '#utils/shouldFlashDroppedItem.js';
 	import sortItems from '#utils/sortItems.js';
 	import ChargeIndicator from '#view/components/ChargeIndicator.svelte';
@@ -128,10 +128,10 @@
 	let itemRulesManagers = new SvelteMap();
 
 	// All charge pools for the actor
-	let allPools = $derived(ChargePoolService.getPools(actor.reactive));
+	let allPools = $derived(getPools(actor.reactive));
 
 	function getItemPools(itemId: string) {
-		return ChargePoolService.getPoolsForItem(actor.reactive, itemId, allPools);
+		return getPoolsForItem(actor.reactive, itemId, allPools);
 	}
 
 	$effect(() => {
